Quick naming fixes
